Remove some unused variables, moved others to RLTrainer
