+++ ml-agents/mlagents/trainers/trainer/rl_trainer.py
@@ -1,5 +1,5 @@
 # # Unity ML-Agents Toolkit
-from typing import Dict
+from typing import Dict, List
 from collections import defaultdict
 import abc
 
@@ -10,6 +10,7 @@ from mlagents.trainers.exception import UnityTrainerException
 from mlagents.trainers.components.reward_signals import RewardSignalResult
 from mlagents_envs.timers import hierarchical_timer
 from mlagents.trainers.agent_processor import AgentManagerQueue
+from mlagents.trainers.trajectory import Trajectory
 
 RewardSignalResults = Dict[str, RewardSignalResult]
 
@@ -31,25 +32,22 @@ class RLTrainer(Trainer):  # pylint: disable=abstract-method
         # collected_rewards is a dictionary from name of reward signal to a dictionary of agent_id to cumulative reward
         # used for reporting only. We always want to report the environment reward to Tensorboard, regardless
         # of what reward signals are actually present.
+        self.cumulative_returns_since_policy_update: List[float] = []
         self.collected_rewards: Dict[str, Dict[str, int]] = {
             "environment": defaultdict(lambda: 0)
         }
         self.update_buffer: AgentBuffer = AgentBuffer()
-        self.episode_steps: Dict[str, int] = defaultdict(lambda: 0)
 
     def end_episode(self) -> None:
         """
         A signal that the Episode has ended. The buffer must be reset.
         Get only called when the academy resets.
         """
-        for agent_id in self.episode_steps:
-            self.episode_steps[agent_id] = 0
         for rewards in self.collected_rewards.values():
             for agent_id in rewards:
                 rewards[agent_id] = 0
 
     def _update_end_episode_stats(self, agent_id: str, optimizer: TFOptimizer) -> None:
-        self.episode_steps[agent_id] = 0
         for name, rewards in self.collected_rewards.items():
             if name == "environment":
                 self.cumulative_returns_since_policy_update.append(
@@ -83,6 +81,48 @@ class RLTrainer(Trainer):  # pylint: disable=abstract-method
         Uses demonstration_buffer to update model.
         """
         pass
+
+    def _increment_step(self, n_steps: int, name_behavior_id: str) -> None:
+        """
+        Increment the step count of the trainer
+        :param n_steps: number of steps to increment the step count by
+        """
+        self.step += n_steps
+        self.next_summary_step = self._get_next_summary_step()
+        p = self.get_policy(name_behavior_id)
+        if p:
+            p.increment_step(n_steps)
+
+    def _get_next_summary_step(self) -> int:
+        """
+        Get the next step count that should result in a summary write.
+        """
+        return self.step + (self.summary_freq - self.step % self.summary_freq)
+
+    def _write_summary(self, step: int) -> None:
+        """
+        Saves training statistics to Tensorboard.
+        """
+        self.stats_reporter.add_stat("Is Training", float(self.should_still_train))
+        self.stats_reporter.write_stats(int(step))
+
+    @abc.abstractmethod
+    def _process_trajectory(self, trajectory: Trajectory) -> None:
+        """
+        Takes a trajectory and processes it, putting it into the update buffer.
+        :param trajectory: The Trajectory tuple containing the steps to be processed.
+        """
+        self._maybe_write_summary(self.get_step + len(trajectory.steps))
+        self._increment_step(len(trajectory.steps), trajectory.behavior_id)
+
+    def _maybe_write_summary(self, step_after_process: int) -> None:
+        """
+        If processing the trajectory will make the step exceed the next summary write,
+        write the summary. This logic ensures summaries are written on the update step and not in between.
+        :param step_after_process: the step count after processing the next trajectory.
+        """
+        if step_after_process >= self.next_summary_step and self.get_step != 0:
+            self._write_summary(self.next_summary_step)
 
     def advance(self) -> None:
         """
